requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
Flask-Caching==2.0.1
Flask-RESTful==0.3.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==5.0.1
celery==5.2.7
gunicorn==20.1.0
pytest==7.3.1
pytest-flask==1.2.0
requests==2.28.2
Werkzeug==2.2.2